Move geomechanics processes to dedicated package
